% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competingrisks.R
\name{crr.formula}
\alias{crr.formula}
\title{Competing risks functions}
\usage{
crr.formula(...)
}
\arguments{
\item{...}{other arguments passed on to \code{cmprsk::crr}}

\item{formula}{MODIFY}

\item{data}{MODIFY}
}
\description{
These functions override the \code{crr} function provided in {cmprsk}
to produce more manageable competing risks model results (i.e. include
model.frame and formula in output) so that results can leverage functions
like \code{broom::tidy} the same way other regression function results do.
}
