% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskybusiness-package.R
\docType{package}
\name{riskybusiness-package}
\alias{riskybusiness}
\alias{riskybusiness-package}
\title{riskybusiness: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/margarethannum/riskybusiness}
  \item Report bugs at \url{https://github.com/margarethannum/riskybusiness/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (YOUR-ORCID-ID)

}
\keyword{internal}
