% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competingrisks.R
\name{crr.formula}
\alias{crr.formula}
\title{Competing risks functions}
\usage{
crr.formula(formula, data, ...)
}
\arguments{
\item{formula}{input formula (e.g Surv(time, event) ~ var1 + var2). event
variable can have multiple levels for use in competing risks.}

\item{data}{MODIFY}

\item{...}{other arguments passed on to \code{cmprsk::crr} (e.g. \code{cencode}, \code{failcode})}
}
\description{
These functions override the \code{crr} function provided in {cmprsk}
to produce more manageable competing risks model results (i.e. include
model.frame and formula in output) so that results can leverage functions
like \code{broom::tidy} the same way other regression function results do.
}
