% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskybiz-package.R
\docType{package}
\name{riskybiz-package}
\alias{riskybiz}
\alias{riskybiz-package}
\title{riskybiz: Handy wrapper for cmprsk package}
\description{
Wrapper functions for to make {cmprsk} package tidy
    and accessible.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/margarethannum/riskybiz}
  \item Report bugs at \url{https://github.com/margarethannum/riskybiz/issues}
}

}
\author{
\strong{Maintainer}: Margie Hannum \email{margie.hannum@gmail.com} (0000-0002-2953-0449)

Authors:
\itemize{
  \item Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (0000-0003-0862-2018)
}

}
\keyword{internal}
