% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskybiz-package.R
\docType{package}
\name{riskybiz-package}
\alias{riskybiz}
\alias{riskybiz-package}
\title{riskybiz: Handy wrapper for cmprsk package}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/margarethannum/riskybusiness}
  \item Report bugs at \url{https://github.com/margarethannum/riskybusiness/issues}
}

}
\author{
\strong{Maintainer}: Margie Hannum \email{margie.hannum@gmail.com} (YOUR-ORCID-ID)

}
\keyword{internal}
